+++ web/src/main/java/com/homeapp/javatraining/exception/user/CannotBlockSelfException.java
@@ -1,0 +1,9 @@
+package com.homeapp.javatraining.exception.user;
+
+import com.homeapp.javatraining.exception.ApiException;
+
+public class CannotBlockSelfException extends ApiException {
+    public CannotBlockSelfException() {
+        super("Cannot block self");
+    }
+}